+++ internal/pkg/runtime/launcher/oci/oci_runc_linux.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2018-2022, Sylabs Inc. All rights reserved.
+// Copyright (c) 2018-2023, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -19,10 +19,8 @@ import (
 	"github.com/sylabs/singularity/internal/pkg/buildcfg"
 	fakerootConfig "github.com/sylabs/singularity/internal/pkg/runtime/engine/fakeroot/config"
 	"github.com/sylabs/singularity/internal/pkg/util/starter"
-	"github.com/sylabs/singularity/pkg/ocibundle/tools"
 	"github.com/sylabs/singularity/pkg/runtime/engine/config"
 	"github.com/sylabs/singularity/pkg/sylog"
-	"github.com/sylabs/singularity/pkg/util/singularityconf"
 )
 
 // Delete deletes container resources
@@ -222,33 +220,23 @@ func Run(ctx context.Context, containerID, bundlePath, pidFile string, systemdCg
 }
 
 // RunWrapped runs a container via the OCI runtime, wrapped with prep / cleanup steps.
-func RunWrapped(ctx context.Context, containerID, bundlePath, pidFile string, systemdCgroups bool) error {
-	// TODO: --oci mode always emulating --compat, which uses --writable-tmpfs.
-	//       Provide a way of disabling this, for a read only rootfs.
-	if err := prepareWriteableTmpfs(bundlePath); err != nil {
-		return err
+func RunWrapped(ctx context.Context, containerID, bundlePath, pidFile string, overlayPaths []string, systemdCgroups bool) error {
+	runFunc := func() error {
+		return Run(ctx, containerID, bundlePath, "", systemdCgroups)
 	}
 
-	err := Run(ctx, containerID, bundlePath, pidFile, systemdCgroups)
-
-	// Cleanup actions log errors, but don't return - so we get as much cleanup done as possible.
-	if err := cleanupWritableTmpfs(bundlePath); err != nil {
-		sylog.Errorf("While cleaning up writable tmpfs: %v", err)
+	if len(overlayPaths) > 0 {
+		return WrapWithOverlays(runFunc, bundlePath, overlayPaths)
 	}
 
-	// Return any error from the actual container payload - preserve exit code.
-	return err
+	return WrapWithWritableTmpFs(runFunc, bundlePath)
 }
 
 // RunWrappedNS reexecs singularity in a user namespace, with supplied uid/gid mapping, calling oci run.
-func RunWrappedNS(ctx context.Context, containerID, bundlePath, pidFile string) error {
+func RunWrappedNS(ctx context.Context, containerID, bundlePath string, overlayPaths []string) error {
 	absBundle, err := filepath.Abs(bundlePath)
 	if err != nil {
 		return fmt.Errorf("failed to determine bundle absolute path: %s", err)
-	}
-
-	if err := os.Chdir(absBundle); err != nil {
-		return fmt.Errorf("failed to change directory to %s: %s", absBundle, err)
 	}
 
 	args := []string{
@@ -256,10 +244,19 @@ func RunWrappedNS(ctx context.Context, containerID, bundlePath, pidFile string) 
 		"oci",
 		"run-wrapped",
 		"-b", absBundle,
-		containerID,
 	}
-	if pidFile != "" {
-		args = append(args, "--pid-file="+pidFile)
+	for _, p := range overlayPaths {
+		absPath, err := absOverlay(p)
+		if err != nil {
+			return fmt.Errorf("could not convert %q to absolute path: %w", p, err)
+		}
+
+		args = append(args, "--overlay", absPath)
+	}
+	args = append(args, containerID)
+
+	if err := os.Chdir(absBundle); err != nil {
+		return fmt.Errorf("failed to change directory to %s: %s", absBundle, err)
 	}
 
 	sylog.Debugf("Calling fakeroot engine to execute %q", strings.Join(args, " "))
@@ -362,18 +359,4 @@ func Update(containerID, cgFile string, systemdCgroups bool) error {
 	cmd.Stdin = os.Stdout
 	sylog.Debugf("Calling %s with args %v", runtimeBin, runtimeArgs)
 	return cmd.Run()
-}
-
-func prepareWriteableTmpfs(bundleDir string) error {
-	sylog.Debugf("Configuring writable tmpfs overlay for %s", bundleDir)
-	c := singularityconf.GetCurrentConfig()
-	if c == nil {
-		return fmt.Errorf("singularity configuration is not initialized")
-	}
-	return tools.CreateOverlayTmpfs(bundleDir, int(c.SessiondirMaxSize))
-}
-
-func cleanupWritableTmpfs(bundleDir string) error {
-	sylog.Debugf("Cleaning up writable tmpfs overlay for %s", bundleDir)
-	return tools.DeleteOverlay(bundleDir)
 }